chore: move EthPrimitives to reth-ethereum-primitives
